requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
python-dotenv==1.0.0
anthropic>=0.40.0
cryptography>=42.0.0
